pyproject.toml
[tool.poetry]
name = "web-hw9"
version = "0.1.0"
description = ""
authors = ["marinasemak <marina.semak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
lxml = "^5.3.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
black = "^24.10.0"
mongoengine = "^0.29.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
